# packages/mtmai/mtmai-0.3.460.tar.gz/mtmai-0.3.460/mtmai/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mtmai",
    "description": "mtmai description(group)",
    "version": "0.3.460"
  },
  "paths": {
...
    "/api/v1/retrieval": {
      "post": {
        "tags": [
          "retrieval"
        ],
        "summary": "Add Content",
        "operationId": "retrieval-add_content",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RagAddContentReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RagAddContentRes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/retrieval/query": {
      "post": {
        "tags": [
          "retrieval"
        ],
        "summary": "Query Doc",
        "operationId": "retrieval-query_doc",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RagRetrievalReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResonse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DocumentPublic": {
        "properties": {
          "collection": {
            "type": "string",
            "title": "Collection"
          },
          "meta": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta",
            "default": {}
          },
          "document": {
            "type": "string",
            "title": "Document"
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "collection",
          "document",
          "id"
        ],
        "title": "DocumentPublic"
      },
      "DocumentResonse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/DocumentPublic"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "data",
          "count"
        ],
        "title": "DocumentResonse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RagAddContentReq": {
        "properties": {
          "collection": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collection"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "collection",
          "content"
        ],
        "title": "RagAddContentReq"
      },
      "RagAddContentRes": {
        "properties": {
          "collection": {
            "type": "string",
            "title": "Collection",
            "default": "default"
          },
          "meta": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta",
            "default": {}
          },
          "document": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8196
              },
              {
                "type": "null"
              }
            ],
            "title": "Document"
          }
        },
        "type": "object",
        "title": "RagAddContentRes"
      },
      "RagRetrievalReq": {
        "properties": {
          "collection": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collection"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "limit": {
            "type": "integer",
            "maximum": 100.0,
            "title": "Limit",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "collection",
          "query"
        ],
        "title": "RagRetrievalReq"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  },
...
}